Add Set Default for Edit and Delete

+++ index.php
@@ -39,8 +39,10 @@
                     echo "<div class='artikel'>";
                     echo "<h2>".$row['title_post']."</h2>";
                     echo "<p>".$row['article']."</p>";
-                    echo"<a class='button-edit' href='edit.php?id=".$row['id_post']."'>Edit</a>";
-                    echo"<a class='button-delete' href='hapus.php?id=".$row['id_post']."'>Hapus</a>";
+                    if($cek == 'login'){
+                        echo"<a class='button-edit' href='edit.php?id=".$row['id_post']."'>Edit</a>";
+                        echo"<a class='button-delete' href='hapus.php?id=".$row['id_post']."'>Hapus</a>";
+                    }
                     echo "</div>";
                     
                 }
